+++ test/parameters.py
@@ -80,8 +80,8 @@ wait_timeout = 120000
 
 
 def iam_principal():
-    # Use 'user principal' or 'instance principal' for production pod, and use
-    # None for minicloud testing.
+    # Use 'user principal', 'instance principal' or 'resource principals' for
+    # production pod, and use None for minicloud testing.
     return None